PIM-10098: move command and handler
